fix(nginx):Make healthcheck port configurable
This fixes the problem that, if there is a second serverblock that uses port 80,
it might be that the liveness and readiness checks fail because they don't work
with the custom server block.
Instead of forcing everyone to set up the correct rules in their serverblock
we'll let those run against port 8080 that's being used for metric scraping anyhow.