fix(prometheus_agent): Update relabel config
This adds two entries to the relabel configs. One of them converts the
"__meta_kubernetes_pod_label_app_kubernetes_io_" labels to something that is usable,
analogous to the relabeling done in promtail.
The other one surfaces the external IP of nodes in the kubelet, which can at points be beneficial to
have available.
The latter one could potentialy present a cardinality problem, but this should be ok, because I would expect
these IPs to change with the hostnames of the nodes.